backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# AI & ML
openai==1.35.0
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2

# Speech & Audio
soundfile==0.12.1

# Document Processing
python-docx==1.1.0
python-multipart==0.0.6

# APIs & Integrations
PyJWT==2.8.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
requests==2.31.0
httpx==0.25.2

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Security & Production
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
sse-starlette==1.8.2